cdk/requirements-dev.txt
pytest==6.2.5
aws-cdk.core
aws-cdk.assertions
aws-cdk.aws_lambda
aws_cdk.aws_s3_deployment
aws_cdk.aws_iam
aws_cdk.aws_dynamodb
aws_cdk.aws_apigatewayv2_integrations
aws_cdk.aws_athena
aws_cdk.aws_sam
aws_cdk.aws_secretsmanager
aws_cdk.aws_ssm
aws_cdk.aws_codecommit
aws_cdk.aws_codepipeline
aws_cdk.aws_codepipeline_actions
aws_cdk.aws_ecs
aws_cdk.aws_ec2
aws_cdk.aws_ecs_patterns
requests
boto3
uvicorn
fastapi>=0.72.0
httpx
starlette